Fix unwanted behavior of CLP when we expect to do dual simplex
CLP is hinted to go for dual simplex but it may still go for primal, giving wrong
results when in strong branching iteration limit is hit and primal has been done.
Moreover, when doing dual simplex, dual solutions may be inaccure at iteration limit.
To fix that, some options are set through Osi to make CLP behave as we expect.