Add few more asserts in dual function implementation
Now CLP dual solutions match the objValue() but there is still something
weird going on, i.e. the evaluate_dual_function() doesn't always match
the optimal value of the B&B even if it is supposed to. This might be
related to:
 - CLP still not providing the correct dual solutions
 - Dual solutions are lost when collected in collect_duals()
 - Dual objective values are not computed correctly in
evaluate_dual_function()

Moreover, Farkas proof still doesn't seem to work. As of now, the proof is
obtained by solving the LP to optimality.