Fix the basis being overwritten during branching
At the beginning of branching, the basis of the current node is kept to warm start cadidates in strong branching.
The basis was kept in some temporary arrays which were used by other functions, causing the basis to be
overwritten. As a quick fix, I just malloc'd memory for the basis, but it is not that clear where to free it.
What if rows/cols are added during branching?