Fix rounding up in evaluate_dual_function() and re-compute the ray in some cases
The global best bound in evaluate_dual_function() must be rounded up according to granularity. As of now, this should work but better to keep an eye on this.
Most of the dual rays from CLP are generated from a dual algorithm and the Farkas proof works on them.  Some rays are generated from primal algorithm (which we don't know why this is called) and the Farkas proof seems not to work only in these cases. While we try to fix it in CLP, I compute a ray by solving a auxiliary LP.